[Fix] 관리자 통계 대시보드 - 최근 3개월 수입 그래프 표시
